task

+++ Part7_Task2/CommandPattern/Point.cs
@@ -8,10 +8,10 @@ namespace CommandPattern
 {
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             this.X = x;
             this.Y = y;
